+++ packages/api/src/controllers/apikey.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Get, Param, Post } from "@nestjs/common";
 import { CreateApikeyDto } from "../core/dtos";
-import { ApikeyUseCases } from "../services/apikey/apikey.use-case";
+import { ApikeyUseCases } from "../services/apikey/apikey.service";
 
 @Controller("apikey")
 export class ApikeyController {
