+++ packages/api/src/controllers/application.controller.ts
@@ -3,7 +3,7 @@ import { ApiOperation } from "@nestjs/swagger";
 import { LoggerService } from "src/core/logger/logger.service";
 import { ExceptionsService } from "src/services/exceptions/exceptions.service";
 import { CreateApplicationDto, UpdateApplicationDto } from "../core/dtos";
-import { ApplicationUseCases, ApplicationFactoryService } from "../use-cases/application";
+import { ApplicationUseCases, ApplicationFactoryService } from "../services/application";
 
 @Controller("application")
 export class ApplicationController {
