+++ packages/api/src/services/application/application.module.ts
@@ -1,8 +1,8 @@
 import { Module } from "@nestjs/common";
 import { LoggerModule } from "src/core/logger/logger.module";
 import { DataServicesModule } from "../data-services/data-services.module";
-import { ApplicationFactoryService } from "./application-factory.service";
-import { ApplicationUseCases } from "./application.use-case";
+import { ApplicationFactoryService } from "./application.factory";
+import { ApplicationUseCases } from "./application.service";
 
 @Module({
   imports: [DataServicesModule],
