+++ packages/api/src/services/property/index.ts
@@ -1,3 +1,3 @@
-export * from "./property-factory.service";
-export * from "./property.use-case";
-export * from "./property-use-cases.module";
+export * from "./property.factory";
+export * from "./property.service";
+export * from "./property.module";
