+++ packages/api/src/services/property/property.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { DataServicesModule } from "../data-services/data-services.module";
-import { PropertyFactoryService } from "./property-factory.service";
-import { PropertyUseCases } from "./property.use-case";
+import { PropertyFactoryService } from "./property.factory";
+import { PropertyUseCases } from "./property.service";
 
 @Module({
   imports: [DataServicesModule],
